Implemented Many to Many relationships
Here, it was mainly about or Pokemon-Categories relationship and our Pokemon-Owners relationship. To represent a many to many relationship, we need to do 2 things: 1. Ensure the join table that will represent the actual relationhips exists, and 2. Integrate that relationship into our Models as properties.

To do that, we first created our 2 new join tables, PokemonCategory and PokemonOwner, and then we created new ICollection<> properties in each entity participating in the many to many relationship.