Created a Controller
The PokemonController uses standard controller architecture, and includes 1 endpoint, an HTTP Get endpoint. With respect to Spring Boot Controllers, The structure is largely the same, including the use of annotations. Although annotations are done with [] instead of @. A new feature is the use of ModelState validation. This is our main way to ensure our Request is good.