Imported and Registered AutoMapper
Automapper is a mapper used to make mapping between Entities and DTOs simple. Similar to MapStruct in java.

First, we imported AutoMapper using NuGet. Then, we created a new mapping profile, similar to creating a mapping class in Java. In that profile's constructor, we then declare a new mapping, between <Pokemon, PokemonDto>. This is similar to creating a new mapping method with MapStruct, PokemonDto pokemonToDto(Pokemon pokemon). Finally, now that we have a fully functional mapping profile, we register it as a service for our app to be able to inject in use. Of course, we do that in our Program.Cs file, using builder.Services.AddAutoMapper.