Expanded controller layer
We expanded our other foundational layer, the controller layer,  by implementing a few new controller endpoins. Like before, we use [] to denote annotations, most importantly to denote the kind of endpoint we are contructing, such as [HttpGet("url/extension")]. To accompany these new methods, we also created a new DTO that corresponsds to our Pokemon Entity

Within our new endpoint methods, we communicate with our repository to get our desired entities, and we use our auto mapper to create a DTO, which typically is the data we expect to sent/recieve to our endpoint consumers.