Implemented DELETE endpoint of Review
We create a new delete method in our repository layer, a delete endpoint in our controller layer, and a service method in our service layer to hold the business logic to delete the entity. This is a relatively simple endpoint to implement, since all that is needed is the entity id.